+++ src/main/java/com/nbs/biz/data/entity/PeerLoginEntity.java
@@ -22,7 +22,7 @@ public class PeerLoginEntity extends BasicEntity {
     private String seckey;
     private String sysuser;
     private Integer ctime;
-    private Integer lmtime;
+    private Long lmtime;
 
     public String getId() {
         return id;
@@ -120,11 +120,11 @@ public class PeerLoginEntity extends BasicEntity {
         this.ctime = ctime;
     }
 
-    public Integer getLmtime() {
+    public Long getLmtime() {
         return lmtime;
     }
 
-    public void setLmtime(Integer lmtime) {
+    public void setLmtime(Long lmtime) {
         this.lmtime = lmtime;
     }
 }
